値が飛ぶ

+++ Core/Inc/MMC5983MA.hpp
@@ -32,9 +32,9 @@
 class MMC5983MA{
 private:
 	struct Offset{
-		int16_t x;
-		int16_t y;
-		int16_t z;
+		int32_t x;
+		int32_t y;
+		int32_t z;
 	};
 	struct Gauss{
 		int32_t x;
